+++ previous-versions/accounting/pkg/models/operations/uploadbillcreditnoteattachment.go
@@ -27,10 +27,13 @@ func (o *UploadBillCreditNoteAttachmentRequestBody) GetRequestBody() string {
 }
 
 type UploadBillCreditNoteAttachmentRequest struct {
-	RequestBody      *UploadBillCreditNoteAttachmentRequestBody `multipartForm:"file" request:"mediaType=multipart/form-data"`
-	BillCreditNoteID string                                     `pathParam:"style=simple,explode=false,name=billCreditNoteId"`
-	CompanyID        string                                     `pathParam:"style=simple,explode=false,name=companyId"`
-	ConnectionID     string                                     `pathParam:"style=simple,explode=false,name=connectionId"`
+	RequestBody *UploadBillCreditNoteAttachmentRequestBody `multipartForm:"file" request:"mediaType=multipart/form-data"`
+	// Unique identifier for a bill credit note.
+	BillCreditNoteID string `pathParam:"style=simple,explode=false,name=billCreditNoteId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
+	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 }
 
 func (o *UploadBillCreditNoteAttachmentRequest) GetRequestBody() *UploadBillCreditNoteAttachmentRequestBody {
@@ -62,11 +65,14 @@ func (o *UploadBillCreditNoteAttachmentRequest) GetConnectionID() string {
 }
 
 type UploadBillCreditNoteAttachmentResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *UploadBillCreditNoteAttachmentResponse) GetContentType() string {