+++ previous-versions/accounting/pkg/models/operations/uploadtransferattachment.go
@@ -27,10 +27,13 @@ func (o *UploadTransferAttachmentRequestBody) GetRequestBody() string {
 }
 
 type UploadTransferAttachmentRequest struct {
-	RequestBody  *UploadTransferAttachmentRequestBody `multipartForm:"file" request:"mediaType=multipart/form-data"`
-	CompanyID    string                               `pathParam:"style=simple,explode=false,name=companyId"`
-	ConnectionID string                               `pathParam:"style=simple,explode=false,name=connectionId"`
-	TransferID   string                               `pathParam:"style=simple,explode=false,name=transferId"`
+	RequestBody *UploadTransferAttachmentRequestBody `multipartForm:"file" request:"mediaType=multipart/form-data"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
+	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
+	// Unique identifier for a transfer.
+	TransferID string `pathParam:"style=simple,explode=false,name=transferId"`
 }
 
 func (o *UploadTransferAttachmentRequest) GetRequestBody() *UploadTransferAttachmentRequestBody {
@@ -62,11 +65,14 @@ func (o *UploadTransferAttachmentRequest) GetTransferID() string {
 }
 
 type UploadTransferAttachmentResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *UploadTransferAttachmentResponse) GetContentType() string {