+++ previous-versions/accounting/pkg/models/shared/profitandlossreport.go
@@ -3,13 +3,14 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/utils"
 	"github.com/ericlagergren/decimal"
 )
 
 type ProfitAndLossReport struct {
-	CostOfSales *ReportLine `json:"costOfSales,omitempty"`
-	Expenses    *ReportLine `json:"expenses,omitempty"`
+	AdditionalProperties map[string]interface{} `additionalProperties:"true" json:"-"`
+	CostOfSales          *ReportLine            `json:"costOfSales,omitempty"`
+	Expenses             *ReportLine            `json:"expenses,omitempty"`
 	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
 	//
 	// ```
@@ -31,16 +32,16 @@ type ProfitAndLossReport struct {
 	// > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
 	FromDate *string `json:"fromDate,omitempty"`
 	// Gross profit of the company in the given date range.
-	GrossProfit types.Decimal `json:"grossProfit"`
-	Income      *ReportLine   `json:"income,omitempty"`
+	GrossProfit *decimal.Big `decimal:"number" json:"grossProfit"`
+	Income      *ReportLine  `json:"income,omitempty"`
 	// Net operating profit of the company in the given date range.
-	NetOperatingProfit types.Decimal `json:"netOperatingProfit"`
+	NetOperatingProfit *decimal.Big `decimal:"number" json:"netOperatingProfit"`
 	// Net other income of the company in the given date range.
-	NetOtherIncome types.Decimal `json:"netOtherIncome"`
+	NetOtherIncome *decimal.Big `decimal:"number" json:"netOtherIncome"`
 	// Net profit of the company in the given date range.
-	NetProfit     types.Decimal `json:"netProfit"`
-	OtherExpenses *ReportLine   `json:"otherExpenses,omitempty"`
-	OtherIncome   *ReportLine   `json:"otherIncome,omitempty"`
+	NetProfit     *decimal.Big `decimal:"number" json:"netProfit"`
+	OtherExpenses *ReportLine  `json:"otherExpenses,omitempty"`
+	OtherIncome   *ReportLine  `json:"otherIncome,omitempty"`
 	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
 	//
 	// ```
@@ -61,6 +62,24 @@ type ProfitAndLossReport struct {
 	// > Not all dates from Codat will contain information about time zones.
 	// > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
 	ToDate *string `json:"toDate,omitempty"`
+}
+
+func (p ProfitAndLossReport) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(p, "", false)
+}
+
+func (p *ProfitAndLossReport) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &p, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *ProfitAndLossReport) GetAdditionalProperties() map[string]interface{} {
+	if o == nil {
+		return nil
+	}
+	return o.AdditionalProperties
 }
 
 func (o *ProfitAndLossReport) GetCostOfSales() *ReportLine {
@@ -84,9 +103,9 @@ func (o *ProfitAndLossReport) GetFromDate() *string {
 	return o.FromDate
 }
 
-func (o *ProfitAndLossReport) GetGrossProfit() types.Decimal {
+func (o *ProfitAndLossReport) GetGrossProfit() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.GrossProfit
 }
@@ -98,23 +117,23 @@ func (o *ProfitAndLossReport) GetIncome() *ReportLine {
 	return o.Income
 }
 
-func (o *ProfitAndLossReport) GetNetOperatingProfit() types.Decimal {
+func (o *ProfitAndLossReport) GetNetOperatingProfit() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.NetOperatingProfit
 }
 
-func (o *ProfitAndLossReport) GetNetOtherIncome() types.Decimal {
+func (o *ProfitAndLossReport) GetNetOtherIncome() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.NetOtherIncome
 }
 
-func (o *ProfitAndLossReport) GetNetProfit() types.Decimal {
+func (o *ProfitAndLossReport) GetNetProfit() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.NetProfit
 }