+++ previous-versions/accounting/pkg/models/shared/propertiestracking1.go
@@ -6,9 +6,11 @@ package shared
 type Propertiestracking1 struct {
 	CategoryRefs []TrackingCategoryRef  `json:"categoryRefs"`
 	CustomerRef  *AccountingCustomerRef `json:"customerRef,omitempty"`
-	IsBilledTo   BilledToType1          `json:"isBilledTo"`
-	IsRebilledTo BilledToType1          `json:"isRebilledTo"`
-	ProjectRef   *ProjectRef            `json:"projectRef,omitempty"`
+	// Defines if the bill or bill credit note is billed/rebilled to a project.
+	IsBilledTo BilledToType1 `json:"isBilledTo"`
+	// Defines if the bill or bill credit note is billed/rebilled to a project.
+	IsRebilledTo BilledToType1 `json:"isRebilledTo"`
+	ProjectRef   *ProjectRef   `json:"projectRef,omitempty"`
 	// Links the current record to the underlying record or data type that created it.
 	//
 	// For example, if a journal entry is generated based on an invoice, this property allows you to connect the journal entry to the underlying invoice in our data model.