+++ previous-versions/accounting/pkg/models/shared/tracking.go
@@ -2,6 +2,10 @@
 
 package shared
 
+import (
+	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/utils"
+)
+
 // TrackingRecordReference - Links the current record to the underlying record or data type that created it.
 //
 // For example, if a journal entry is generated based on an invoice, this property allows you to connect the journal entry to the underlying invoice in our data model.
@@ -27,11 +31,30 @@ func (o *TrackingRecordReference) GetID() *string {
 }
 
 type Tracking struct {
+	AdditionalProperties map[string]interface{} `additionalProperties:"true" json:"-"`
 	// Links the current record to the underlying record or data type that created it.
 	//
 	// For example, if a journal entry is generated based on an invoice, this property allows you to connect the journal entry to the underlying invoice in our data model.
 	InvoiceTo  *TrackingRecordReference `json:"invoiceTo,omitempty"`
 	RecordRefs []InvoiceTo              `json:"recordRefs"`
+}
+
+func (t Tracking) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(t, "", false)
+}
+
+func (t *Tracking) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &t, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *Tracking) GetAdditionalProperties() map[string]interface{} {
+	if o == nil {
+		return nil
+	}
+	return o.AdditionalProperties
 }
 
 func (o *Tracking) GetInvoiceTo() *TrackingRecordReference {